rename: addEvent callback 분리 -> controller
