+++ src/client/js/components/Header/Search/components/SearchInput.js
@@ -1,6 +1,10 @@
 import Component from "../../../../core/Component";
 import { createExtendsRelation } from "../../../../core/oop-utils";
-import { delay, request } from "../../../../core/utils";
+import {
+  handleInputFocusIn,
+  handleInputFocusOut,
+  handleKeyupWithFocus,
+} from "../controllers/searchInput";
 
 function SearchInput(...params) {
   Component.call(this, ...params);
@@ -12,56 +16,15 @@ SearchInput.prototype.setup = function () {
     inputData: "",
   };
 };
+
 SearchInput.prototype.setEvent = function () {
-  this.addEvent("focusout", "input[type='text']", () => {
-    const { searchSuggestion, searchRecent } = this.$props;
-    searchSuggestion.setState({ display: "none" });
-    searchRecent.setState({ display: "none" });
-  });
-  this.addEvent("focusin", "input[type='text']", () => {
-    const { inputData } = this.state;
-    const { searchRecent, searchSuggestion } = this.$props;
-    if (inputData) {
-      searchSuggestion.setState({ display: "flex" });
-      searchRecent.setState({ display: "none" });
-    } else {
-      searchSuggestion.setState({ display: "none" });
-      searchRecent.setState({ display: "flex" });
-    }
-  });
-  this.addEvent("keyup", "input[type='text']", ({ target }) => {
-    const { searchSuggestion, searchRecent } = this.$props;
-    if (target.value) {
-      searchRecent.setState({ display: "none" });
-    } else {
-      searchRecent.setState({ display: "flex" });
-    }
-    this.state.inputData = target.value;
-    /* 5초 뒤에도 같은 값인지 확인 하기위한 변수 */
-    const curValue = target.value;
-    delay(500).then(async () => {
-      if (this.state.inputData === curValue) {
-        const requestOptions = {
-          query: {
-            keyword: curValue,
-          },
-        };
-        const { results: suggestionDatas } = await request(
-          "search/autoComplete",
-          requestOptions
-        );
-        if (suggestionDatas?.length) {
-          searchSuggestion.setState({
-            suggestionDatas,
-            word: curValue,
-            display: "flex",
-          });
-        } else {
-          searchSuggestion.setState({ display: "none" });
-        }
-      }
-    });
-  });
+  this.addEvent(
+    "focusout",
+    "input[type='text']",
+    handleInputFocusOut.bind(this)
+  );
+  this.addEvent("focusin", "input[type='text']", handleInputFocusIn.bind(this));
+  this.addEvent("keyup", "input[type='text']", handleKeyupWithFocus.bind(this));
 };
 
 SearchInput.prototype.mount = function () {
